chat/requirements.txt
fastapi
uvicorn
PyJWT
httpx
python-multipart
requests
uvicorn[standard]
websockets
aiohttp